Adding method comments to AiController and Agent, about done with controller
